Refactor so de::Error is always available even if the "binder" feature isn't used
